+++ lib/sidekiq/logging.rb
@@ -18,6 +18,12 @@ module Sidekiq
       end
     end
 
+    class WithoutTimestamp < Pretty
+      def call(severity, time, program_name, message)
+        "#{::Process.pid} TID-#{Thread.current.object_id.to_s(36)}#{context} #{severity}: #{message}\n"
+      end
+    end
+
     def self.with_context(msg)
       Thread.current[:sidekiq_context] ||= []
       Thread.current[:sidekiq_context] << msg
@@ -30,7 +36,7 @@ module Sidekiq
       oldlogger = defined?(@logger) ? @logger : nil
       @logger = Logger.new(log_target)
       @logger.level = Logger::INFO
-      @logger.formatter = Pretty.new
+      @logger.formatter = ENV['DYNO'] ? WithoutTimestamp.new : Pretty.new
       oldlogger.close if oldlogger && !$TESTING # don't want to close testing's STDOUT logging
       @logger
     end